No need to undo test rename
